+++ databricks_dbapi/hive.py
@@ -14,15 +14,12 @@ threadsafety = hive.threadsafety
 paramstyle = hive.paramstyle
 
 
-def connect(
-    host, port=443, database="default", cluster=None, http_path=None, token=None, user=None, password=None, org=0
-):
-    """Create a Hive DBAPI connection to an interactive Databricks cluster.
+def connect(host, port=443, database="default", http_path=None, token=None, user=None, password=None):
+    """Create a pyhive-driven DBAPI connection to an interactive Databricks cluster.
 
     Create a DBAPI connection to a Databricks cluster, which can be used to generate
-    DBAPI cursor(s). Provide either a ``cluster`` OR an ``http_path`` from the cluster's
-    JDBC/ODBC connection details. If using Azure, ``http_path`` is required. On
-    instantiation, ``http_path`` is prioritized over ``cluster`` and ``org``.
+    DBAPI cursor(s). Provide an ``http_path`` from the cluster's JDBC/ODBC connection
+    details.
 
     For authentication, provide either a ``token`` OR both a ``user`` and ``password``.
     Token authentication is strongly recommended over passwords.
@@ -33,15 +30,11 @@ def connect(
     :param str host: the server hostname from the cluster's JDBC/ODBC connection page.
     :param int port: the port number from the cluster's JDBC/ODBC connection page.
     :param str database: the database to use
-    :param str cluster: the cluster unique name or alias. Not required if passing
-        ``http_path``
     :param str http_path: the HTTP Path as shown in the cluster's JDBC/ODBC connection
-        page. Required if using Azure platform.
+        page.
     :param str token: a Databricks API token.
     :param str user: a Databricks user name.
     :param str password: the corresponding Databricks user's password.
-    :param str org: the org id associated with the Databricks workspace (E2). Required
-        for E2 if not passing ``http_path``. Not required if passing ``http_path``.
     """
     if token is not None:
         auth = "token:%s" % token
@@ -58,8 +51,6 @@ def connect(
 
     if http_path is not None:
         url = "https://%s:%s/%s" % (host, port, http_path)
-    elif cluster is not None:
-        url = "https://%s:%s/sql/protocolv1/o/%s/%s" % (host, port, org, cluster)
     else:
         raise ValueError("Missing arguments. Must provide either cluster or http_path.")